+++ Sources/View/Channel/MessageCell/OpenChannel/SBUOpenChannelFileMessageCell.swift
@@ -17,8 +17,6 @@ open class SBUOpenChannelFileMessageCell: SBUOpenChannelContentBaseMessageCell {
         return self.message as? SBDFileMessage
     }
     
-    
-    // MARK: - Private property
     public lazy var baseFileContentView: SBUBaseFileContentView = {
         let fileView = SBUBaseFileContentView()
         return fileView