+++ Sources/View/Channel/Reaction/SBUMessageReactionView.swift
@@ -11,40 +11,40 @@ import SendBirdSDK
 
 
 /// Emoji reaction box
-class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+open class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
 
-    lazy var collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-    let layout: UICollectionViewFlowLayout = SBUCollectionViewFlowLayout()
+    public lazy var collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+    public let layout: UICollectionViewFlowLayout = SBUCollectionViewFlowLayout()
 
-    var emojiList: [SBDEmoji] = []
-    var reactions: [SBDReaction] = []
-    var maxWidth: CGFloat = SBUConstant.messageCellMaxWidth
+    public var emojiList: [SBDEmoji] = []
+    public var reactions: [SBDReaction] = []
+    public var maxWidth: CGFloat = SBUConstant.messageCellMaxWidth
 
     @SBUThemeWrapper(theme: SBUTheme.componentTheme)
-    var theme: SBUComponentTheme
+    public var theme: SBUComponentTheme
 
     var emojiTapHandler: ((_ emojiKey: String) -> Void)? = nil
     var moreEmojiTapHandler: (() -> Void)? = nil
     var emojiLongPressHandler: ((_ emojiKey: String) -> Void)? = nil
 
-    private var collectionViewHeightConstraint: NSLayoutConstraint!
-    private var collectionViewMinWidthContraint: NSLayoutConstraint!
+    public private(set) var collectionViewHeightConstraint: NSLayoutConstraint!
+    public private(set) var collectionViewMinWidthContraint: NSLayoutConstraint!
 
-    private let collectionViewInsets = UIEdgeInsets(top: 8, left: 8, bottom: 8, right: 8)
+    public let collectionViewInsets = UIEdgeInsets(top: 8, left: 8, bottom: 8, right: 8)
     
     lazy var moreEmojiTapRecognizer = UITapGestureRecognizer(
         target: self,
         action: #selector(self.onTapMoreEmoji(sender:))
     )
         
-    init() {
+    public init() {
         super.init(frame: .zero)
         self.setupViews()
         self.setupAutolayout()
         setupStyles()
     }
 
-    override init(frame: CGRect) {
+    public override init(frame: CGRect) {
         super.init(frame: frame)
         self.setupViews()
         self.setupAutolayout()
@@ -52,11 +52,11 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
     }
 
     @available(*, unavailable, renamed: "MessageReactionView()")
-    required init?(coder: NSCoder) {
+    public required init?(coder: NSCoder) {
         super.init(coder: coder)
     }
 
-    func setupViews() {
+    open func setupViews() {
         self.layout.sectionInset = UIEdgeInsets(top: 8, left: 8, bottom: 8, right: 4)
         self.layout.minimumInteritemSpacing = 4
         self.layout.minimumLineSpacing = 4
@@ -77,7 +77,7 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
         self.addSubview(self.collectionView)
     }
 
-    func setupAutolayout() {
+    open func setupAutolayout() {
         self.collectionView.setConstraint(from: self, left: 0, right: 0, top: 0, bottom: 0)
 
         self.collectionViewHeightConstraint = self.collectionView
@@ -89,7 +89,7 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
         self.collectionViewMinWidthContraint.isActive = true
     }
 
-    func setupStyles() {
+    open func setupStyles() {
         self.clipsToBounds = true
         self.backgroundColor = theme.reactionBoxBackgroundColor
         self.layer.cornerRadius = 16
@@ -97,12 +97,12 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
         self.layer.borderColor = theme.reactionBoxBorderLineColor.cgColor
     }
 
-    func hasMoreEmoji(at indexPath: IndexPath) -> Bool {
+    public func hasMoreEmoji(at indexPath: IndexPath) -> Bool {
         return self.reactions.count < emojiList.count &&
             self.reactions.count == indexPath.row
     }
     
-    func configure(maxWidth: CGFloat, useReaction: Bool, reactions: [SBDReaction]) {
+    open func configure(maxWidth: CGFloat, useReaction: Bool, reactions: [SBDReaction]) {
         guard useReaction, !reactions.isEmpty else {
             self.collectionViewMinWidthContraint.isActive = false
             self.isHidden = true
@@ -135,13 +135,15 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
         self.setNeedsLayout()
     }
 
-    private func getCellSize(count: Int) -> CGSize {
+    /// The default value is `CGSize(width: 54, height: 30)`; if `count` is zero, the width is 36.
+    open func getCellSize(count: Int) -> CGSize {
         return CGSize(width: count > 0 ? 54 : 36, height: 30)
     }
 
     // MARK: - Action
 
-    @objc func onTapMoreEmoji(sender: UITapGestureRecognizer) {
+    @objc
+    open func onTapMoreEmoji(sender: UITapGestureRecognizer) {
         let indexPath = IndexPath(row: reactions.count, section: 0)
         collectionView.selectItem(at: indexPath, animated: true, scrollPosition: .init())
         collectionView.deselectItem(at: indexPath, animated: true)
@@ -150,11 +152,11 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
 
 
     // MARK: - UICollectionView relations
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
+    open func numberOfSections(in collectionView: UICollectionView) -> Int {
         return 1
     }
 
-    func collectionView(_ collectionView: UICollectionView,
+    open func collectionView(_ collectionView: UICollectionView,
                         numberOfItemsInSection section: Int) -> Int {
         guard !reactions.isEmpty else { return 0 }
 
@@ -165,7 +167,7 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
         }
     }
 
-    func collectionView(_ collectionView: UICollectionView,
+    open func collectionView(_ collectionView: UICollectionView,
                         cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         
         guard let cell = collectionView.dequeueReusableCell(
@@ -204,7 +206,7 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
         return cell
     }
 
-    func collectionView(_ collectionView: UICollectionView,
+    open func collectionView(_ collectionView: UICollectionView,
                         didSelectItemAt indexPath: IndexPath) {
         
         guard !self.hasMoreEmoji(at: indexPath) else { return }
@@ -213,7 +215,7 @@ class SBUMessageReactionView: UIView, UICollectionViewDelegate, UICollectionView
         self.emojiTapHandler?(reaction.key)
     }
 
-    func collectionView(_ collectionView: UICollectionView,
+    open func collectionView(_ collectionView: UICollectionView,
                         layout collectionViewLayout: UICollectionViewLayout,
                         sizeForItemAt indexPath: IndexPath) -> CGSize {