+++ Sources/View/MessageSearch/SBUMessageSearchResultCell.swift
@@ -11,7 +11,6 @@ import SendBirdSDK
 
 @objcMembers
 open class SBUMessageSearchResultCell: UITableViewCell {
-    
     // MARK: - Properties (View)
     public private(set) lazy var coverImage = SBUCoverImageView()
     public private(set) lazy var titleLabel: UILabel = {
@@ -260,9 +259,13 @@ open class SBUMessageSearchResultCell: UITableViewCell {
             self.messageLabel.text = message.message
         }
         
-        let createdAt = Date.lastUpdatedTime(baseTimestamp: message.createdAt)
-        self.createdAtLabel.text = createdAt ??
-            Date.sbu_from(message.createdAt).sbu_toString(format: .hhmma)
+        let createdAt = Date.lastUpdatedTime(
+            baseTimestamp: message.createdAt,
+            timeFormat: SBUDateFormatSet.MessageSearch.sentTimeFormat
+        )
+        self.createdAtLabel.text = createdAt ?? Date
+            .sbu_from(message.createdAt)
+            .sbu_toString(dateFormat: SBUDateFormatSet.MessageSearch.sentTimeFormat)
     }
     
     /// Sets file message icon depending on the message's file type using `SBUUtils.getFileType(by: fileMessage)`.